+++ src/itson.nim
@@ -1,9 +1,10 @@
-from std/tables import Table, `[]=`, `$`
+from std/tables import Table, `[]=`, `$`, `[]`
 from std/json import parseJson, to, `$`
 import std/jsonutils
 from std/os import fileExists
-from std/times import getTime, toUnix
+from std/httpclient import newAsyncHttpClient, get, Http200, code, close
 
+import std/times
 import std/locks
 
 import pkg/prologue
@@ -15,16 +16,36 @@ type
 proc getCache(file: string): Cache =
   ## Retrieves the cache from a json file
   if fileExists file:
-    let node = file.readFile.parseJson.to Cache
-    echo node
+    result = file.readFile.parseJson.to Cache
 proc setCache(file: string; data: Cache) =
   ## Saves the cache to a json file
   file.writeFile data.toJson.`$`
 
-proc itson*(sitesJson, cacheJson: string) =
+proc checkSiteOn(site: string): Future[bool] {.async.} =
+  let
+    client = newAsyncHttpClient()
+    res = await client.get site
+  result = res.code == Http200
+  close client
+
+proc isOn(file, site: string; delay: int): Future[bool] {.async.} =
+  ## Check if site is on and save it to cache
+  var cache = getCache file
+  let now = getTime()
+
+  if not cache.hasKey site:
+    cache[site] = (false, 0'i64)
+
+  result = cache[site].online
+
+  if cache[site].time.fromUnix + delay.milliseconds <= now:
+    result = await checkSiteOn site
+    cache[site] = (result, now.toUnix)
+    file.setCache cache
+
+proc itsOn*(sitesJson, cacheJson: string; delay: int) =
   var sitesLock: Lock
   let sites {.guard: sitesLock.} = sitesJson.readFile.parseJson.to Sites
-
 
   proc check(ctx: Context) {.async.} =
     ## Checks if website is online
@@ -34,12 +55,9 @@ proc itson*(sitesJson, cacheJson: string) =
       withLock sitesLock:
         if sites.hasKey id:
           site = sites[id]
-    var cache = cacheJson.getCache
-    cache[site] = (true, getTime().toUnix)
-    echo cache
-    cacheJson.setCache cache
 
-    resp "<h1>" & site & "</h1>"
+    let on = await cacheJson.isOn(site, delay)
+    resp if on: "1" else: "0"
     
   initLock sitesLock
 
@@ -52,4 +70,4 @@ proc itson*(sitesJson, cacheJson: string) =
 
 when isMainModule:
   import pkg/cligen
-  dispatch itson
+  dispatch itsOn