+++ changelog.md
@@ -3,3 +3,6 @@
 ## Version 0.1.0 (Aug 9, 2022)
 
 - Init
+- Added custom url code: `GET /{code}` to check a site without the URL
+- Added custom time cache to avoid request overflow
+